requirements/common.txt
django-extensions~=3.0.5
django~=3.1
djangorestframework~=3.11.1
elastic-apm~=5.8.1
psycopg2-binary~=2.8.5
psycopg2~=2.8.5